move flag.Parse from init()
